eduvision-ai/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
opencv-python>=4.8.1.78
numpy>=1.26.0
pillow>=11.0.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.2
pytest>=7.4.3
tensorflow>=2.17.0
deepface>=0.0.89